PBR working correctly again
